final release

+++ src/main/java/com/javarush/kaminsky/entity/Animal.java
@@ -1,7 +1,7 @@
 package com.javarush.kaminsky.entity;
 
 public abstract class Animal extends Being {
-    public Animal(int health, int quantity, int speed, double food, String view) {
+    public Animal(double health, int quantity, int speed, double food, String view) {
         super(health, quantity, speed, food, view);
     }
