+++ src/main/java/com/javarush/kaminsky/entity/carnivores/Bear.java
@@ -1,7 +1,7 @@
 package com.javarush.kaminsky.entity.carnivores;
 
 public class Bear extends Carnivore {
-    public Bear(int health, int quantity, int speed, double food, String view) {
+    public Bear(double health, int quantity, int speed, double food, String view) {
         super(health, quantity, speed, food, view);
     }
