+++ src/main/java/com/javarush/kaminsky/entity/carnivores/Wolf.java
@@ -1,8 +1,7 @@
 package com.javarush.kaminsky.entity.carnivores;
 
 public class Wolf extends Carnivore {
-    public Wolf(int health, int quantity, int speed, double food, String view) {
+    public Wolf(double health, int quantity, int speed, double food, String view) {
         super(health, quantity, speed, food, view);
     }
-
 }
