+++ src/main/java/com/javarush/kaminsky/entity/herbivores/Herbivore.java
@@ -3,7 +3,7 @@ package com.javarush.kaminsky.entity.herbivores;
 import com.javarush.kaminsky.entity.Animal;
 
 public abstract class Herbivore extends Animal {
-    public Herbivore(int health, int quantity, int speed, double food, String view) {
+    public Herbivore(double health, int quantity, int speed, double food, String view) {
         super(health, quantity, speed, food, view);
     }
